Исправил обязательные + частично можно лучше
