+++ src/services/reducers/index.js
@@ -1,12 +1,10 @@
 import { combineReducers } from "redux";
-import {
-  ingredientsReducer,
-  constructorReducer,
-  viewedIngredientReducer,
-  orderElementReducer,
-  modalReducer,
-  tabBarReducer,
-} from "./app";
+import { ingredientsReducer } from "./ingredients";
+import { constructorReducer } from "./constructor";
+import { viewedIngredientReducer } from "./viewed-ingredient";
+import { orderElementReducer } from "./order-element";
+import { modalReducer } from "./modal";
+import { tabBarReducer } from "./tab-bar";
 
 export const rootReducer = combineReducers({
   ingredients: ingredientsReducer,
